fix(audit): P0→P3 — distances km, RPCs HallOfFame, RLS, tests manquants
P0: conversion distances mètres→km dans seed.sql et activity_feed JSONB
P0: VIEW shoes_with_km corrigée (suppression /1000), schema.sql mis à jour
P0: 3 RPCs SECURITY DEFINER (get_running_leaderboard, get_workout_leaderboard, get_hall_of_fame_records)
P1: policies RLS UPDATE ajoutées sur workout_sessions, running_sessions, community_challenges
P1: suppression appel mort increment_shoe_km dans running.service
P2: weather_temp parseFloat→parseInt, deleteRecord silent failure, Community.tsx cast any
P3: fix isOverdue timezone (comparaison dates sans heure)
P3: tests services badges, profile, auth (27 fichiers, 367 tests ✅)
P3: tests pages RunSession et Community

+++ src/pages/RunSession.tsx
@@ -115,7 +115,7 @@ export function RunSessionPage() {
         elevation_loss: elevationLoss ? parseInt(elevationLoss) : undefined,
         avg_heart_rate: avgHeartRate ? parseInt(avgHeartRate) : undefined,
         max_heart_rate: maxHeartRate ? parseInt(maxHeartRate) : undefined,
-        weather_temp: weatherTemp ? parseFloat(weatherTemp) : undefined,
+        weather_temp: weatherTemp ? parseInt(weatherTemp) : undefined,
         weather_condition: weatherCondition || undefined,
         shoe_id: shoeId || undefined,
         feedback: feedback || undefined,
